# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.GPU:
    description: GPUs device from where telemetry collected
    properties:
      device:
        type: string
      gpu_id:
        type: integer
      hostname:
        type: string
      uuid:
        type: string
    type: object
  model.GPUResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/model.GPU'
        type: array
      total:
        type: integer
    type: object
  model.JSONBMap:
    additionalProperties:
      type: string
    type: object
  model.Telemetry:
    description: Telemetry data collected from GPUs
    properties:
      Hostname:
        type: string
      container:
        type: string
      created_at:
        type: string
      device:
        type: string
      gpu_id:
        type: integer
      id:
        type: integer
      labels_raw:
        $ref: '#/definitions/model.JSONBMap'
      metric_name:
        type: string
      modelName:
        type: string
      namespace:
        type: string
      pod:
        type: string
      timestamp:
        type: string
      uuid:
        type: string
      value:
        type: number
    type: object
  model.TelemetryResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/model.Telemetry'
        type: array
      total:
        type: integer
    type: object
info:
  contact: {}
  description: API for querying GPU telemetry stored in TimescaleDB
  title: Telemetry API
  version: "1.0"
paths:
  /gpus:
    get:
      description: Returns list of unique GPU IDs available in telemetry data.
      parameters:
      - default: 10
        description: Limit number of records
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GPUResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List GPUs
      tags:
      - Telemetry
  /gpus/{gpu_id}/metrics:
    get:
      description: Returns metrics data for a given GPU ID and metrics name.
      parameters:
      - description: GPU ID
        in: path
        name: gpu_id
        required: true
        type: string
      - description: Metric name
        enum:
        - DCGM_FI_DEV_DEC_UTIL
        - DCGM_FI_DEV_ENC_UTIL
        - DCGM_FI_DEV_FB_FREE
        - DCGM_FI_DEV_FB_USED
        - DCGM_FI_DEV_GPU_TEMP
        - DCGM_FI_DEV_GPU_UTIL
        - DCGM_FI_DEV_MEM_CLOCK
        - DCGM_FI_DEV_MEM_COPY_UTIL
        - DCGM_FI_DEV_POWER_USAGE
        - DCGM_FI_DEV_SM_CLOCK
        in: query
        name: metric
        required: true
        type: string
      - description: Aggregation function
        enum:
        - min
        - max
        - avg
        - sum
        in: query
        name: aggregation
        required: true
        type: string
      - description: 'Start time (sample: 2025-08-28T08:46:58Z)'
        in: query
        name: start_time
        type: string
      - description: 'End time (sample: 2025-08-28T09:46:58Z)'
        in: query
        name: end_time
        type: string
      - default: 10
        description: Limit number of records
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TelemetryResponse'
      summary: Get metrics for a GPU
      tags:
      - Telemetry
  /gpus/{gpu_id}/telemetry:
    get:
      description: Returns telemetry data for a given GPU ID.
      parameters:
      - description: GPU ID
        in: path
        name: gpu_id
        required: true
        type: string
      - description: 'Start time (sample: 2025-08-28T08:46:58Z)'
        in: query
        name: start_time
        type: string
      - description: 'End time (sample: 2025-08-28T09:46:58Z)'
        in: query
        name: end_time
        type: string
      - default: 10
        description: Limit number of records
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TelemetryResponse'
      summary: Get telemetry for a GPU
      tags:
      - Telemetry
swagger: "2.0"